refactor(gateway): split platform tests
